Create BluePrint for door and add openable component
Create BluePrints for the door and visualize the opening.
To make this posible a new openable C++ component is used.
This component generates events for open/close witch are used by BluePrint.
The openable component based on an Actor with a switchable component to generate the events for open/close.