Geting the Wallswitch working
Make BluePrints for GameMode and DefaultPawn to enable imput handling and enable components for the DefaultPawn. Create a new Manipulator component, and set it on the BluePrinted DefaultPawn. With the help of the InputBounding and the Manipulator component, it is now posible to manipulate the WallSwitch. The WallSwitch must be set to Physics... to get the LineTrace of the Manipulator component working.